// pharmacy-management-frontend/src/app/model/DrugDTO.ts
export interface DrugDTO {
  drugId: string;
   drugCode: string;
   drugName: string;
   activeElement: string;
    drugAmount: string;
    unit: string;
    conversionUnit: string;
    conversionRate: string;
    wholesaleProfitRate: string;
    retailProfitRate: string;
    drugFaculty: string;
    manufacturer: string;
    origin: string;
    note: string;
    drugSideEffect: string;
    flag: string;
    drugGroup: string;
    importPrice: string;
    discountRate: string;
    vat: string;
    wholesalePrice: string;
    retailPrice: string;
}